+++ .idea/encodings.xml
@@ -3,5 +3,8 @@
   <component name="Encoding">
     <file url="file://$PROJECT_DIR$/clinic-management-system-/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/clinic-management-system-/src/main/resources" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/clinic-management-system/src/main/resources" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/src/main/resources" charset="UTF-8" />
   </component>
 </project>
