+++ clinic-management-system-/src/main/java/clinicamed/controller/Controller.java
@@ -1,4 +1,4 @@
-package controller;
+package clinicamed.controller;
 
 import javafx.fxml.Initializable;
 import java.net.URL;
