+++ clinic-management-system-/src/main/java/module-info.java
@@ -2,8 +2,9 @@ module com.clinic.project.medclinic {
     requires javafx.controls;
     requires javafx.fxml;
 
-    exports com.clinic.project.medclinic;
-    exports controller;
-    opens controller to javafx.fxml;
-    opens com.clinic.project.medclinic to javafx.fxml;
+    exports clinicamed.mainapp;
+    exports clinicamed.controller;
+
+    opens clinicamed.mainapp to javafx.fxml;
+    opens clinicamed.controller to javafx.fxml;
 }
