+++ pom.xml
@@ -41,6 +41,12 @@
     </dependencies>
 
     <build>
+        <sourceDirectory>clinic-management-system-/src/main/java</sourceDirectory>
+        <resources>
+            <resource>
+                <directory>clinic-management-system-/src/main/resources</directory>
+            </resource>
+        </resources>
         <plugins>
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
@@ -60,7 +66,7 @@
                         <!-- Default configuration for running with: mvn clean javafx:run -->
                         <id>default-cli</id>
                         <configuration>
-                            <mainClass>com.clinic.project.medclinic/appllication.MainApplication
+                            <mainClass>clinicamed.mainapp.MainApplication
                             </mainClass>
                             <launcher>app</launcher>
                             <jlinkZipName>app</jlinkZipName>
